dwarf: Handle filename in DWARF5 properly
DWARF5 changed the way it handles file name info and it seems to treat
index 0 at the file table as the default file name.  But libdw has been
handing it differently and ignoring the 0 index.

This makes srcline related feasture not working on binaries built with
clang which supports DWARF5.  Check it when dwarf_decl_file() return
NULL but we have AT_decl_file.  We can use the first file name then.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>