python: Build symbol table for python functions
It needs to map python function name to an address so that uftrace can
deal with it like C functions.  As lowest address won't be used by
native programs, just use simple counter (from 1) to save symbol
address.  The symbol table will be written as "python.fake.sym" in the
data directory.

But it also needs to handle multiprocessing.  Each process will see its
own copy of the symbol table, but we want to maintain a single table
for output.  The current code uses a shared memory and atomic operations
to update the symbol table contents.

The header area in the symbole table has a counter and offset, and the
rest will be used like a symbol file.  At the end, it'll overwrite the
header with comments.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>