tests: Fix t225_dynamic_size.py by detecting function size
t225_dynamic_size.py is a test for -Z/--size-filter option and it
currently use the size threshold with a magic number 100.

But it sometimes fails as follows because there is no gurantee that the other
function sizes are greater than 100.

  $ ./runtest.py -c clang -Os 225
  Start 1 tests without worker pool
  Compiler                  clang
  Test case                 pg fi fp
  ------------------------: Os Os Os
  225 dynamic_size        : NG NG NG

This patch fixes those unintended failures in t225_dynamic_size.py by
precisely detecting the function size of 'small'.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>