tests: Make self.p_flag empty when testing dynamic tracing
self.p_flag is set to '-fpatchable-function-entry=N' for '-e' option in
runtest.py.

This makes conflicts when testing dynamic tracing itself so we should
suppress this default '-P .' option when testing dynamic tracing.

In addition, this patch changes the cflags order to avoid '-P .' is
being applied at the end, which has the highest priority.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>