tests: Fix args tests in clang build
Clang optimizer is smart enough not to pass arguments when given
arguments are not used inside function body.

Many args tests in clang builds fail due to this optimization.

  062 arg_char            : OK OK OK OK OK SK SK SK SK SK OK OK OK OK OK  OK NG NG NG NG SK SK SK SK SK OK NG NG NG NG
  084 arg_mixed           : OK OK OK OK OK SK SK SK SK SK OK OK OK OK OK  OK NG NG NG NG SK SK SK SK SK OK NG NG NG NG
  085 arg_reg             : OK OK OK OK OK SK SK SK SK SK OK OK OK OK OK  OK NG NG NG NG SK SK SK SK SK OK NG NG NG NG
  201 arg_dwarf1          : OK OK OK OK OK SK SK SK SK SK OK OK OK OK OK  OK NG NG NG NG SK SK SK SK SK OK NG NG NG NG
  254 arg_dwarf5          : OK OK OK OK OK SK SK SK SK SK OK OK OK OK OK  OK NG NG NG NG SK SK SK SK SK OK NG NG NG NG

This patch creates global variables and makes all the given arguments
store its values to the corresponding global variables.

This patch fixes the above failures in clang build and the following run
command shows all the above are fixed.

  $ ./runtest.py '062|084|085|201|254'

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>