uftrace: Don't patch all functions by default
Originally, when the user asks to only unpatch functions, uftrace would
patch all other functions by default. This is counter-intuitive,
especially when using the agent to unpatch at runtime. The user doesn't
expect functions to be patched when they only unpatch funtions.

This commit removes this behavior, and requires the user to explicitly
define functions to patch.

Co-authored-by: Gabriel-Andrew Pollo-Guilbert <gabrielpolloguilbert@gmail.com>
Signed-off-by: Clément Guidi <cguidi@ciena.com>