Catch IndexError from self.sort in test/runtest.py
When packaging uftrace for Fedora using the Copr build service,
I enabled running the test suite for checking the build.

I found that on some CentOS 8 build targets, a few tests caused
test/runtest.py to crash because they referenced an variable[1]
without first checking if variable[1] can be indexed.

Catch this exception in runtest.py and (hopefully) handle
it correctly as a diff in the test case result.