misc: Add more prototypes enums for socket
The following socket() shows some incorrect enum values for now.

  int socket(int domain, int type, int protocol);

The incorrect output looks like this.

  socket(AF_NETLINK, SOCK_PACKET|SOCK_DCCP|SOCK_SEQPACKET|SOCK_RDM|SOCK_RAW|SOCK_DGRAM|SOCK_STREAM+0x807e4, 15) = 7;

It's because the 'type' field can accept SOCK_CLOEXEC and SOCK_NONBLOCK,
but current prototypes.h doesn't have them.

So this patch adds them then the output looks correct as follows.

  socket(AF_NETLINK, SOCK_CLOEXEC|SOCK_NONBLOCK|SOCK_RAW, 15) = 7;

This can be seen as follows.

  $ uftrace -la -C socket lsusb
      ...
  # DURATION     TID     FUNCTION
              [ 18635] | libusb_init() {
              [ 18635] |   udev_monitor_new_from_netlink() {
     5.682 us [ 18635] |     socket(AF_NETLINK, SOCK_CLOEXEC|SOCK_NONBLOCK|SOCK_RAW, 15) = 7;
    13.064 us [ 18635] |   } /* udev_monitor_new_from_netlink */
     6.142 ms [ 18635] | } /* libusb_init */

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>