cmd/recv.c: Migrate from deprecated gethostbyname() to getaddrinfo()
The Fedora review-tool runs rpmlint on the binary packages and warned
that the uftrace tool was using gethostbyname. Switching to getaddrinfo
helps to get accepted, also for later uftrace distro packaging reviews:

POSIX.1-2008 removes the specifications of gethostbyname(), gethostbyaddr(),
and h_errno, recommending the use of getaddrinfo(3) and getnameinfo(3) instead:

They return a list of addrs or hostnames, are thread-safe and optionally
allow us to also use IPv6. This 1st step continues with IPv4 and only
still tries to connect to the primary, 1st address (for a small commit).

Signed-off-by: Bernhard Kaindl <contact@bernhard.kaindl.dev>