test: Use negative index for sorting report output
In some cases, (self) time calculation can result in 0 and current
output just omit the column.  But it can confuse the test sort code for
uftrace report which assumes specific index of the output.  Using
negative index can be a solution since it mostly want to get the last
few columns.  Get rid of some comments at the end as well.

Although we now have a try/except logic to prevent crashes, it'd be nice
if the sort code can handle that more naturally.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>