dynamic: Batch function unpatching process
Synchronization requires sending signals to threads inside the process.
This is performed for every symbols, which means a lot of signal can be
sent. This has a major performance impact. This commit batches the
initial and final steps of the patching process so we only need to send
one signal per thread for every batch.

Co-authored-by: Gabriel-Andrew Pollo-Guilbert <gabrielpolloguilbert@gmail.com>
Signed-off-by: Clément Guidi <cguidi@ciena.com>