mcount: Use flag to indicate if target is running
Use a global flag to indicate the state of the target. When the target
is not running, tasks such as dynamic patching can be performed with
less constraints.

If libmcount.so is dynamically injected (not implemented yet), the
'mcount_target_running' flag indicates that libmcount has to be
initialized in a running target.

Signed-off-by: Clément Guidi <cguidi@ciena.com>