mcount: Deep copy 'mcount_triggers' rbtree
Perform a deep copy of 'mcount_triggers' so the agent can implement a
RCU-like algorithm to modify filters and triggers at runtime.

The deep copy process doesn't use rbtree primitives, as it takes
advantage of the fact that the original tree is (supposed to be) valid,
and the copy doesn't modify the structure.

Signed-off-by: Clément Guidi <cguidi@ciena.com>