python: Add the directory of the script to import path
When it runs a script from a subdirectory, it should add the dir
to the import path so that it can load other modules in the same
directory.

  $ uftrace mydir/script.py

Normally the python interpreter would do the job, but it'd see
the uftrace.py loader instead so we need to do it manually.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>