dynamic: x86_64: Do not allow ret insn for patch target
Since 'ret' instruction can be a target of another jump instruction,
it is not safe to patch and move the instruction somewhere else.

For example, the following instruction pattern must not be patched
because the branch target of 'jne' instruction at ef0af8 is the address
of the 'ret' instruction at ef0ac8.

  0000000000ef0ac0 <_ZN2v88internal4Heap20CreateFillerObjectAtEmiNS0_20ClearFreedMemoryModeE>:
    ef0ac0:       f3 0f 1e fa             endbr64
    ef0ac4:       85 d2                   test   %edx,%edx
    ef0ac6:       75 08                   jne    ef0ad0 <_ZN2v88internal4Heap20CreateFillerObjectAtEmiNS0_20ClearFreedMemoryModeE+0x10>
    ef0ac8:       c3                      ret
        ...
    ef0af8:       75 ce                   jne    ef0ac8 <_ZN2v88internal4Heap20CreateFillerObjectAtEmiNS0_20ClearFreedMemoryModeE+0x8>

If the 'ret' is patched, then it gets crashed with SIGILL and the
prologue of the function becomes as follows.

  $ cgdb --args uftrace record -P v8::internal::Heap::CreateFillerObjectAt --no-libcall ./node -e ''
        ...
  Thread 2.1 "node" received signal SIGILL, Illegal instruction.
  0x00005555562f0aca in v8::internal::Heap::CreateFillerObjectAt(unsigned long, int, v8::internal::ClearFreedMemoryMode) ()

   1├>Dump of assembler code for function _ZN2v88internal4Heap20CreateFillerObjectAtEmiNS0_20ClearFreedMemoryModeE:
   2│    0x00005555562f0ac0 <+0>:     endbr64
   3│    0x00005555562f0ac4 <+4>:     call   0x5555578eeff0
         ...
  16│    0x00005555562f0af8 <+56>:    jne    0x5555562f0ac8 <_ZN2v88internal4Heap20CreateFillerObjectAtEmiNS0_20ClearFreedMemoryModeE+8>

The 'call' instruction at 5555562f0ac4 is the patched one, but the
target of 'jne' is pointing the address in the middle of the 'call'
instruction, which causes the SIGILL.

To avoid this problem, we rejects the 'ret' instruction and more
candidates to make it safer.

Fixed: #1692
Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>