python: Treat modules under the same dir as main
Often large python programs are divided into separate modules and the
main script just loads them and calls the entry routines.  In this case
the separate modules should be treated as main module so that the
internal functions and traced fully.

We assume those modules are placed in the same directory with the main
script.  So compare the filename of those modules and check if they have
the same prefix (the directory path).

Signed-off-by: Namhyung Kim <namhyung@gmail.com>