mcount-dynamic: Use MAP_FIXED_NOREPLACE for trampoline
By trying to allocate the extra page for a trampoline with MAP_FIXED,
Uftrace could end up overlapping already mapped pages, which are
discarded.  Typically, if data was there with rw- protection, now the
data is undefined and has rwx protection.

Since Linux 4.17, the flag MAP_FIXED_NOREPLACE can be used to ensure
that the fixed location is mapped atomically, preventing any clobbering.

For earlier kernel, the call to mmap(2) will fall back to a random
address, by not honoring the MAP_FIXED request at all.  So callers
should ensure that the returned address matches the requested one.

Signed-off-by: Olivier Dion <odion@efficios.com>