tui: Add Ctrl+u/d, Ctrl+e/y keys to move half page and scroll line
problem:
tui needs to move the cursor to the top and bottom to see the previous
or next page code.

solution:
Add the keys to move the page by plcaing the cursor on the result
that you are interested in.

Those are functions in vim and I implemented it.
you can move and search the result more flexibly.

Fixed: #1802
Signed-off-by: kang-hyuck <kang-hyuck@naver.com>