mcount:Temporary fix find_got for no-plt binary
Issue #1777 reported that no-plt binary is handled as plt-enable
binary because some functions are still left in plt section.
The patch fixes this bug by comparing the name of self with
a list of known functions.

However, current `mcount_arch_hook_no_plt` make segfault on x86
now, so I comment it out and print the debug message if the binary
was built with `-fno-plt` option.

Signed-off-by: ChoKyuWon <kyuwoncho18@gmail.com>