replay: Revise the logic of setting NEEDS_PAREN to correctly add parentheses
The full demangler __cxa_demangle() may or may not add parentheses
at the end of symbol names, so NEEDS_PAREN was introduced.

Previously, NEEDS_PAREN was set when the last character of the symbol
name was not ')'. Now, it is also set when the last few characters are
"operator()", to handle functors and lambdas in C++.

Related to #1827

Signed-off-by: Yufeng Jin <jinyufeng2000@gmail.com>