tests: Add two unit tests for dump command
The uftrace dump uses the data in two ways.  One is to dump each file
contents separately without checking time ordering.  This is for the raw
dump to check consistency in the file.  The other is to dump the whole
contents ordered by time like uftrace replay.  This is to generate time
chart like chrome and so on.

So let's add a test case for each usage to verify its functionality (at
the very minimum) and to check any bugs or leaks.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>