mcount: Update debug message for GOT updates
When it sets up PLT hooking, it'd overwrite GOT entries for each library
functions at offset + 3.  But the plt entry/exit functions print the
original symbol table index.  This can be confusing so let's make it
explicit by printing "+ 3" part.

Before:
  plthook: overwrite GOT[415] to 0x7f460c9db500 (__cyg_profile_func_exit)

After:
  plthook: overwrite GOT[412 + 3] to 0x7f460c9db500 (__cyg_profile_func_exit)

So not it matches the later messages like:

  plthook: [idx:  412] enter 56304c2019f0: __cyg_profile_func_exit@plt (mod: 7f6a2a51b180)

Signed-off-by: Namhyung Kim <namhyung@gmail.com>