misc: Enhance and rename wget-pr script
I improved the recently added wget-pr.sh script.
It's really useful to get the patch file by pull request number.
However, some systems (e.g. CentOS 7 or 8, fedora 33 or 34)
do not provide wget unless install manually.

This commit improves the script to get the patch with curl if wget is not available.
This also renames the file to a more appropriate name.

Signed-off-by: Kang Minchul <tegongkang@gmail.com>