configure: Check /proc/sys/kernel/perf_event_paranoid
In order to get kernel information with uftrace without any problem,
I think /proc/sys/kernel/perf_event_paranoid should be 2 or less.
But many distros nowadays set this value to more than 3 as default.

I wanted to inform users about this in the docs, but it's
so general that it's unclear which docs to leave this information in.

So this commit suggests checking this information
when configuring uftrace.

Closes: #1424

Signed-off-by: Kang Minchul <tegongkang@gmail.com>