utils: Add DTRAP macro and use it in pr_err
We don't have much information when uftrace hits pr_err, but it just
shows the given error message.

Instead, it'd be useful if we can attach gdb inside pr_err so that we
can see its backtrace and so on.

This patch adds DTRAP, which enables TRAP only when debug mode is on.

We can add DTRAP just before pr_err goes to exit().

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>