live: Allow --list-event without argument
The option itself doesn't take an argument.  But current option parsing
logic didn't allow live (or default) command running without an argument
(for target executable).  This is good for normal cases but --list-event
can run without argument and it'll just try to print kernel events if it
is root.  See #923.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>