record: Use pipe instead of eventfd
It uses an eventfd to sync parent and child to setup environment
before starting the actual tracing.  But it seems eventfd cannot handle
the case where the parent existed early.  In that case child will be
stuck on the eventfd waiting for someone to write.

The pipe will reject such case (IOW read will return an error) because
it has clear role for each fd while an eventfd can be used for read and
write so it won't reject read without writers.

Fixed: #1866
Signed-off-by: Namhyung Kim <namhyung@gmail.com>