arch/arm: Do not exit when libmcount cannot find symbol
On ARM (32-bit), mcount_arch_parent_location() needs to look at the
address of the current function to determine if it's in THUMB2 or ARM
ISA.  But in the test code, it fails to find one as it has no symbol
table.  So it should allow the case and just return the original address
in the location hoping that it would have a proper address.

Furthermore, --estimate-return doesn't require the actual return
address.  Thus it should not exit even if it couldn't find a symbol for
child function.

Fixed: #1867
Signed-off-by: Namhyung Kim <namhyung@gmail.com>