build: Add -mno-sse for x86_64 build
It was reported that using -mno-sse2 alone can crash clang compiler when
the code it compiles accesses floating-point or vector types.  It seems
adding -mno-sse fix the problem.  I thought -mno-sse2 implies it but
that doesn't seem like the case and caused a problem.

Actually recent versions of clang have -mgeneral-regs-only option and we
should prefer it.  But let's keep the fix minimal for now.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>