mcount: Compile out float types in script support
The libmcount is built not to generate FP/SIMD access due to performance
reasons.  But script support needs to access them anyway.  I think it
won't work correctly in libmcount.  We might restrict accessing FP args
in the scripts which runs on uftrace record.

Define LIBMCOUNT macro to rule out the FP code from the libmcount.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>