mcount: Do not keep FD for message channel
uftrace and libmcount communicate using a pipe in the data directory.
Historically it used to have an unnamed pipe but it's using a FIFO which
means it doesn't need to keep the file descriptor open.

Instead it can open the file whenever it needs to send a message. While
this would bring more overheads, a micro benchmark showed it's very
small. The upside is that it can work with daemon programs which close
all file descriptors at the beginning.

Signed-off-by: Rongsong Shen <rshen@shenrs.eu>