build: Do not use libelf in the python extension
The python extension module is to trace python functions and it doesn't
need to deal with ELF files in general.  The only place needed is to get
the address of libmcount functions.  As it's the minimal functionality,
we can force to use the symbol-rawelf code by removing HAVE_LIBELF
definition in the PYTHON_CFLAGS.  It'd make it lighter without further
dependency.

But this requires separate handling the object files for the python
module.  Let's use a different object file suffix (.oy) to apply the
compiler flags separately.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>