test: Update lib_dlopen2 test case (t200)
It has a couple problems.  First, the filter should be set for the main
function, there's no function called 'a'.  And the 'creat' function is
not the syscall so it should not fix up it to match creat64 in libc.

Finally, it seems GCC optimized out the Parent construction as it does
nothing.  clang seems to keep it, so let's handle it in the fixup().

The failed output looked like below:
    t200_lib_dlopen2: diff result of g++ -pg -O1 -g -fno-ipa-sra
    --- expect	2024-03-02 12:33:32.381859085 -0800
    +++ result	2024-03-02 12:33:32.381859085 -0800
    @@ -3,7 +3,5 @@
        dlsym();
    -   creat64() {
    -     Child::Child() {
    -       Parent::Parent();
    -     } /* Child::Child */
    -   } /* creat64 */
    +   creat() {
    +     Child::Child();
    +   } /* creat */
        Child::func();

Signed-off-by: Namhyung Kim <namhyung@gmail.com>