test: Make flamegraph test (t102) less flaky
Unlike other dump tests, it's unstable since it requires to create
sample for the given period.  When system load is light, it'd work as
expected, but if not, the usleep() would be blocked more than the given
timeout and there's no way to control it.

Let's just ignore the number of samples, we want correct callstacks
only.  The failed output looked like:

    t102_dump_flamegraph: diff result of gcc -pg -O2  -fno-ipa-sra
    --- expect	2024-03-02 12:24:41.321268477 -0800
    +++ result	2024-03-02 12:24:41.321268477 -0800
    @@ -1 +1 @@
    -main;foo;bar;usleep 1
    +main;foo;bar;usleep 2

Signed-off-by: Namhyung Kim <namhyung@gmail.com>