test: Update trace script test case (t220)
This is to trace shell scripts with traceable binaries.  By default it'd
trace PLT functions in the shell binary.  And it executes test binary
(t-abc) and let's limit the output just for the test binary.  Remove the
filter on 'fork' function for simplicity.

Also it needs to trace dynamically if the binary is built with -fpatchable-
function-entry flag.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>