kernel: Update the pattern of syscall trace functions
Sometimes it has functions like syscall_trace_enter.isra.0 depending
on the compiler optimization so it's hard to write the exact name.
Let's use a pattern to handle multiple functions with different suffix
properly.

Note that kernel internally uses glob patterns, and uftrace accepts
regex by checking list of filter functions before passing the input to
the kernel.  As it can add the name directly to the kernel here, using
the glob pattern would be easier.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>