test: Fix kernel dump test (t103)
I've noticed failures on the kernel test with the chrome tracing format
dump because of JSON load failure and a format error.  It seems to use
the double quote consistently in the metadata.  Also add missing TID
part in the process/thread name.

Also chrome_sort() logic should not filter out functions with "__"
because the kernel syscall routines start with it.  In the kernel test,
it needs to check the kernel function.  So let's remove it in the sort
logic and filter the functions (usually before 'main' in the user
program) by using -F filter explicitly.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>