report: Fix output field length
The new stdv fields add one more space after the percent sign and break
alignment.

Also fix the TUI output as it should include the '*' sign for the
currently selected sort key.  Change the field length to the actual
number it prints and move the two spaces between fields to the
win_display_report() to make it explicit.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>