info: Add utc_offset info for external usage
In some cases, there is a requirement to get the time offset between UTC
time and our timestamp from the given clock source, CLOCK_MONOTONIC by
default.

This patch provides an additional info to provide UTC time offset in the
output of uftrace info.

This number can be used to make time adjustment easier for external
tools such as tracecompass[1].

Closes: #1916

[1] https://eclipse.dev/tracecompass
Requested-by: Matthew Khouzam <matthew.khouzam@gmail.com>
Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>