dump: Skip kernel function records for unrelated tasks
I found that sometimes the kernel ftrace data contains record from
unrelated tasks even if it has the task filter.  In that case it cannot
get a valid task pointer and result in a segfault due to the NULL
dereference.  Maybe it's a bug in the kernel, but we should be able to
run uftrace without a problem.

Reported-by: George Xanthakis <georgexanthak@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>