misc: Add checkout-pr.sh script for easier checkout
It usually takes some time to get patches for review but it'd be useful
to have a script for that.

The 'checkout-pr.sh' script accepts a single parameter to get the pull
request ID.
It fetches from the source repo and checkouts to the fetched head.

The example usage is as follows.

  $ ./misc/checkout-pr.sh 1974
    % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                   Dload  Upload   Total   Spent    Left  Speed
  100 18502  100 18502    0     0  39899      0 --:--:-- --:--:-- --:--:-- 39875
  remote: Enumerating objects: 1, done.
  remote: Counting objects: 100% (1/1), done.
  remote: Total 1 (delta 0), reused 1 (delta 0), pack-reused 0 (from 0)
  Unpacking objects: 100% (1/1), 354 bytes | 354.00 KiB/s, done.
  From https://github.com/GabrielKimm/uftrace
   * branch              fix-tui-non-utf8-locale -> FETCH_HEAD
  Switched to and reset branch 'pull/1974'

Now you're ready to review the pull request 1974 where can be found at
https://github.com/namhyung/uftrace/pull/1974.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>