arch/x86: Fix segfault after mcount() on temporary registers
It seems some aggresively optimizing compilers use temporary registers
like r10 across function calls.  The -fipa-ra might enable this but I
didn't think it'll do that over function calls it cannot see the
definition.  Anyway it happens and we should be able to deal with it
like we did in dynamic.S.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>