mcount: Rename to mcount_{rstack,auto}_rehook()
The rstack restore and reset functions (and their auto counterparts)
are to temporarily change the return addresses in the stack.  But I
found it confusing what restore and reset mean when I look at the code
some time later.  So let's clarify the names.

Basically libmcount hooks the return address (unless -e is used) to
redirect it to mcount library when the function returns.  Now restore
means to change the (hooked) return address to its original address
(as it it's not called uftrace), and rehook means to change the
*restored* return address to redirect to the libmcount again.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>