utils: Identify graph entry nodes to skip
Identify if the current graph entry must be skipped druing `dump`, which
is done during the add_graph_entry() procedure.

Basic idea is to find recursive calls, but the first recursive call
cannot be skipped. For example, if the call trace shows:

main - fib - fib - fib

then the first (fib - fib) must be present but the second must be
skipped, making the graph look like:

main - fib - fib

To separate recursive calls from the first recursive call, determine
if the node's name is the same with curr and additionally check if
curr->parent's name is the same as well.

Signed-off-by: Daon Park <deeopark39@gmail.com>