utils: Transfer skipped node's nr_calls
If a node is skipped due to being a recursive call, then it's nr_calls
must be correctly appended to the first recursive call. During the
`add_graph_entry()` procedure, it looks for the source of the first
recursive call and saves it to recursive_src. Next, if the current
graph entry node is skipped, then the nr_calls is added to the
add_calls_tgt.

Signed-off-by: Daon Park <deeopark39@gmail.com>