misc: Rename benchmark functions intuitively
It check average and minimum execution time of each function.  But it's
not clear what the functions mean.  Let's rename them to show one is a
leaf function and others are parent and child.  I've added 'nested_'
prefix to have them together because I want to sort the result by name
of the function.

Also remove the schedule function from the output.

  $ misc/bench.sh
  # uftrace bench
      Self avg    Self min  Function
    ==========  ==========  ====================
      0.028 us    0.024 us  leaf
      0.028 us    0.024 us  nested_child
      0.083 us    0.065 us  nested_parent

Signed-off-by: Namhyung Kim <namhyung@gmail.com>