graph, tui(G): Add total/self avg, max, min fields to -f option in cmds,tui
This patch adds support for the following fields in the graph -f:
- total-avg, total-max, total-min
- self-avg, self-max, self-min

These fields show the average, maximum, and minimum time based on each function call time.

The simple usage is as follows.

$ uftrace graph -f total-avg,total-max,total-min
$ uftrace graph -f self-avg,self-max,self-min

In tui simply press 'f' and check the corresponding fields

Then you will see the functions, similar to those shown in the report fields.

Fixed: #1451

Signed-off-by: Jun Yeong Kim <junyeonggim5@gmail.com>