plthook: Use mcount_arch_ops.entry to skip PLTHOOK
Some internal functions should not go through the PLT.  Instead of using
weak aliases, let's use the entry address in mcount_arch_ops.

But the cygprof functions are not saved in the table, it needs to handle
them separately.  This will fix the following error when trying to build
with LTO.

    LINK     libmcount/libmcount.so
  libmcount/plthook.c:93:1: error: function ‘uftrace___cyg_profile_func_exit’ redeclared as variable
     93 | ALIAS_DECL(__cyg_profile_func_exit);
        | ^
  libmcount/mcount.c:1994:1: note: previously declared here
   1994 | UFTRACE_ALIAS(__cyg_profile_func_exit);
        | ^
  libmcount/plthook.c:92:1: error: function ‘uftrace___cyg_profile_func_enter’ redeclared as variable
     92 | ALIAS_DECL(__cyg_profile_func_enter);
        | ^
  libmcount/mcount.c:1988:1: note: previously declared here
   1988 | UFTRACE_ALIAS(__cyg_profile_func_enter);

Signed-off-by: Namhyung Kim <namhyung@gmail.com>