test: Sleep 100ms to wait for recv started
During the runtime test, the recv test sometimes failed due to the
connection-refused error.  It was because uftrace record tried to
connect to the recv server before it prepared the socket.

When I added a message in uftrace recv, I can see this:

    uftrace: /home/namhyung/project/uftrace/cmds/recv.c:125:setup_client_socket
     ERROR: socket connect failed (host: localhost, port: 45679): Connection refused
    uftrace: /home/namhyung/project/uftrace/cmds/record.c:2240:do_child_exec
     ERROR: waiting for parent failed: Invalid argument
    server socket created: port 45679

Add 100ms of sleep after uftrace recv to mitigate the failure.
It's not a perfect solution but it seems to work well for me.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>