filter: Add TRIGGER_FL_CONDITION flag
It's to check if the given trigger has a filter condition.  It was
checked by cond.idx being non-zero before, but it'd be consistent to use
a flag for that.

Also changed the value of TRIGGER_FL_CLEAR.  It should be fine as the
value is only used in memory and not saved to a data file.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>