dump:  Add UTC offset to Chrome trace format
This patch applies the UTC offset to timestamps in Chrome trace
output. When enabled, the timestamps represent actual UTC time rather
than monotonic time, making it easier to correlate trace events with
external time-based data or logs.

The UTC offset information is already available in the trace metadata
(added in commit fae85a8), and this patch leverages that data to
adjust timestamps when the option is specified.

Signed-off-by: Kaveh Shahedi <kavehshahedi@gmail.com>